+++ lizard_progress/models.py
@@ -8,6 +8,7 @@ import logging
 
 from django.contrib.gis.db import models
 from django.contrib.auth.models import User
+from django.core.urlresolvers import reverse
 
 from lizard_map.models import JSONField
 
@@ -15,13 +16,14 @@ import lizard_progress.specifics
 
 logger = logging.getLogger(__name__)
 
+
 def has_access(user, project, contractor=None):
     """Test whether user has access to this project (showing data of
     this contractor.
 
     Should probably be implemented using lizard-security but I lack
     the time to figure that out right now."""
-    
+
     if user.is_anonymous():
         # Not logged in
         return False
@@ -93,7 +95,7 @@ class Location(models.Model):
     information = JSONField(null=True, blank=True)
 
     objects = models.GeoManager()
-    
+
     class Meta:
         # IDs are unique within a project
         unique_together = (("unique_id", "project"))
@@ -111,7 +113,8 @@ class MeasurementType(models.Model):
 
     icon_missing = models.CharField(max_length=50, null=True, blank=True)
     icon_complete = models.CharField(max_length=50, null=True, blank=True)
-    global_icon_complete = models.CharField(max_length=50, null=True, blank=True)
+    global_icon_complete = models.CharField(
+        max_length=50, null=True, blank=True)
 
     def __unicode__(self):
         return u"Type '%s' in %s" % (self.name, self.project.name)
@@ -122,25 +125,53 @@ class ScheduledMeasurement(models.Model):
     contractor = models.ForeignKey(Contractor)
     measurement_type = models.ForeignKey(MeasurementType)
     location = models.ForeignKey(Location)
+    timestamp = models.DateTimeField(auto_now=True)
 
     complete = models.BooleanField()
 
-    @property
-    def measurement(self):
-        try:
-            return Measurement.objects.get(scheduled=self)
-        except Measurement.DoesNotExist:
-            return None
-
 
 class Measurement(models.Model):
+    """Although most ScheduledMeasurements will have only a single
+    associated measurements, some will have more because there is only
+    a single file associated with a measurements.
+
+    E.g. HDSR's Dwarsprofielen project has a measurement type "fotos"
+    that calls for two photos, one of the left bank and one of the
+    right bank. That must be represented by two measurements, since
+    there are two different files. However, there is only one
+    scheduled measurement."""
+
     scheduled = models.ForeignKey(ScheduledMeasurement)
 
-    # Dict with the data, or image filename/URL, etc.
+    # Dict with the data.
     # Site decides what to store.
     data = JSONField(null=True)
 
+    # The date as retrieved from the measurement data, not the date of
+    # uploading.
     date = models.DateTimeField(null=True, blank=True)
+
+    # Any available geometry data in the uploaded measurement.
     the_geom = models.PointField(null=True, blank=True, srid=SRID)
 
+    # This is the filename of the uploaded file after it was moved
+    # into lizard-progress' archive. Set by the upload view.
+    filename = models.CharField(max_length=1000)
+
+    # Auto-changes, most likely set in stone the time the file is
+    # uploaded.
+    timestamp = models.DateTimeField(auto_now=True)
+
     objects = models.GeoManager()
+
+    @property
+    def url(self):
+        """Return the URL to the uploaded file that contained this
+        measurement."""
+        
+        sm = self.scheduled
+        return reverse('lizard_progress_filedownload', kwargs={
+                    'project_slug': sm.project.slug,
+                    'contractor_slug': sm.contractor.slug,
+                    'measurement_type_slug': sm.measurement_type.slug,
+                    'filename': os.path.basename(self.filename)})